requirements.txt
streamlit
pandas
pytesseract
Pillow
sqlite3
